backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Athlthieu17 <hieult.hust170703@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.104.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.0.3"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.22"
alembic = "^1.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
